+++ storage/src/backend/registry.rs
@@ -534,7 +534,7 @@ impl BlobReader for RegistryReader {
             .map_err(BackendError::Registry)
     }
 
-    fn prefetch_blob_data_range(&self, _ra_offset: u32, _ra_size: u32) -> BackendResult<()> {
+    fn prefetch_blob_data_range(&self, _ra_offset: u64, _ra_size: u64) -> BackendResult<()> {
         Err(BackendError::Unsupported(
             "Registry backend does not support prefetch as per on-disk blob entries".to_string(),
         ))
